+++ app/models/events.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import uuid
 from datetime import datetime
+from decimal import Decimal
 from typing import TYPE_CHECKING
 
 from sqlalchemy import DateTime, Enum, ForeignKey, Integer, Numeric, Text, UUID, VARCHAR
@@ -31,7 +32,7 @@ class Event(Base, TimeStamp):
     description: Mapped[str] = mapped_column(Text, nullable=False)
     venue: Mapped[str] = mapped_column(VARCHAR(255), nullable=False)
     starts_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), nullable=False)
-    ticket_price: Mapped[float] = mapped_column(Numeric(10, 2), nullable=False)
+    ticket_price: Mapped[Decimal] = mapped_column(Numeric(10, 2), nullable=False)
     total_tickets: Mapped[int] = mapped_column(Integer, nullable=False)
     status: Mapped[EventStatus] = mapped_column(Enum(EventStatus, name="event_status"), nullable=False)
     organizer_id: Mapped[uuid.UUID] = mapped_column(ForeignKey("users.id", ondelete="CASCADE"), nullable=False)
@@ -60,4 +61,3 @@ class Event(Base, TimeStamp):
         back_populates="events",
         overlaps="event_tags,event,tag",
     )
-
